added more things but not tested

+++ DMS/app/src/main/java/com/example/kbb12/dms/FitnessInfo/IFitnessInfo.java
@@ -8,5 +8,5 @@ import android.content.Context;
 
 public interface IFitnessInfo {
 
-    public double getCalCount(Context context);
+    public int getCalCount(Context context);
 }
